% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ct.R
\name{ct_run}
\alias{ct_run}
\title{perform predictions}
\usage{
ct_run(sce, model = "Immune_All_Low.pkl", majority_voting = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment instance, assumed to have counts as assay component}

\item{model}{character(1) name of a celltypist model with .pkl suffix}

\item{majority_voting}{logical(1), if TRUE, a super-clustering procedure is run
and voting across clustering outcomes is used to assign cell type}
}
\value{
A SingleCellExperiment with additional colData columns `predicted_labels` and
`conf_score` from celltypist
}
\description{
perform predictions
}
\examples{
z = scRNAseq::ZilionisLungData()
r = ct_run(z)
names(r)
}
