% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typist.R
\name{get_ct}
\alias{get_ct}
\title{Simple demonstration, not basilisk-compliant, but hands back python reference}
\usage{
get_ct()
}
\description{
Simple demonstration, not basilisk-compliant, but hands back python reference
}
