% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_desc.R
\name{model_desc}
\alias{model_desc}
\title{acquire the python object corresponding to a celltypist model}
\usage{
model_desc(mname = "Human_Lung_Atlas.pkl")
}
\arguments{
\item{mname}{character(1) model name with .pkl suffix, or path to a pickle file holding a model}
}
\description{
acquire the python object corresponding to a celltypist model
}
\note{
This hands back a live reference to a python object, which does
not respect the typical basilisk protocol.  A safer version would extract
atomic R data and package it as the return value.
}
\examples{
pa = system.file(file.path("pickle", "Human_PF_Lung.pkl"), package="typistR")
model_desc(pa)
}
